Move parser classes into dedicated file
