+++ src/python/spectrseqtools/preprocessing/preprocessing.py
@@ -13,6 +13,7 @@ from clr_loader import get_mono
 from spectrseqtools.common import set_output_path
 from spectrseqtools.enums import AveragineBackbone
 from spectrseqtools.masses import initialize_nucleotide_df
+from spectrseqtools.parsers import PreprocessingOptions
 from spectrseqtools.preprocessing.deconvolution import (
     DeconvolutionParameters,
     DeisotopedPeakList,
@@ -28,7 +29,7 @@ rt = get_mono()
 class Preprocessor:
     """Class for preprocessing of raw MS data."""
 
-    def __init__(self, options) -> None:
+    def __init__(self, options: PreprocessingOptions) -> None:
         self.alphabet = initialize_nucleotide_df(input_path=options.alphabet)
         self.tolerance = options.tolerance
         self.singleton_boundaries = SingletonBoundaries.from_alphabet(
